fix most compilation errors with vc++ (there are still a problem with eina_log and output of the log messages: STDERR_FILENO, STDOUT_FILENO and isatty())
fix some warnings and header files inclusion

update vcproj files


SVN revision: 47384

+++ legacy/eina/src/include/eina_log.h
@@ -19,6 +19,7 @@
 #ifndef EINA_LOG_H_
 #define EINA_LOG_H_
 
+#include <stdlib.h>
 #include <stdarg.h>
 
 #include "eina_types.h"
@@ -222,7 +223,7 @@ struct _Eina_Log_Domain
    int level;              /**< Max level to log */
    const char *domain_str; /**< Formatted string with color to print */
    const char *name;       /**< Domain name */
-   unsigned int namelen;   /**< strlen(name) */
+   size_t namelen;   /**< strlen(name) */
 
    /* Private */
    Eina_Bool deleted:1;    /**< Flags deletion of domain, a free slot */
