+++ legacy/eina/src/modules/mp/fixed_bitmap/eina_fixed_bitmap.c
@@ -83,8 +83,8 @@ _eina_fixed_cmp_key(const Eina_Rbtree *node, const void *key, __UNUSED__ int len
 {
    const void *a = node;
    const void *b = key;
-   int delta;
-   int limit;
+   ssize_t delta;
+   ssize_t limit;
 
    limit = sizeof (Eina_Fixed_Bitmap_Pool) + mp->item_size * 32;
    delta = (char *)a - (char *)b;
@@ -150,7 +150,7 @@ eina_fixed_bitmap_free(void *data, void *ptr)
    Eina_Fixed_Bitmap_Pool *pool;
    void *a;
    Eina_Bool push_front = EINA_FALSE;
-   int delta;
+   ssize_t delta;
 
    pool = (Eina_Fixed_Bitmap_Pool*) eina_rbtree_inline_lookup(mp->lookup,
 							      ptr, 0,
